Finalização dos testes Repositorio de Order

+++ DonaLaura/DonaLaura.Domain/Features/Orders/IOrderRepository.cs
@@ -14,6 +14,6 @@ namespace DonaLaura.Domain.Features.Orders
         IEnumerable<Order> GetAll();
         void Delete(Order order);
 
-        IEnumerable<Order> GetByProduct(int idProduct);
+        IEnumerable<Order> GetByProduct(long idProduct);
     }
 }
